Switched from int.Parse to double.Parse in LemonadeStandOwner.SetPrice()

+++ LemonadeStand/LemonadeStandOwner.cs
@@ -120,7 +120,7 @@ namespace LemonadeStand
         public void SetPrice()
         {
             Console.WriteLine("Enter new price per cup: ");
-            lemonadeCupPrice = int.Parse(Console.ReadLine());
+            lemonadeCupPrice = double.Parse(Console.ReadLine());
             Console.WriteLine($"Price set to {lemonadeCupPrice}");
             Console.ReadLine();
         }
